Fix requireToken()
Right now method chaining doesn't work, e.g. you can't do the following:

$user = JWTAuth::parseToken()->toUser();

JWTAuth::parseToken() will set the token property in JWTAuth class to the token in the header. If we chain ->toUser() the requireToken() method will not throw an exception as there is a token set to the object, but in the end we just pass in the $token (which is false) to setToken(). Therefore, throwing an TokenInvalidException('Wrong number of segments').

The change should enable that we can chain these methods.