Fix issues with storage provider tests
Using the "demeter chain" syntax wasn't actually checking whether `tags()` was getting called, because Mockery shortcuts the check when it knows what the output would be (and only verifies the final method in the chain). Instead we separate the method checks, as well as enforcing they are called with `once()` and duplicating them to have separate tests for the tagged and untagged formats.