Adjust cache taggability checks
Existing checks for whether a cache supported `tags()` were quietly failing because it was handled by magical `__call` in L5.1.27 and lower. Instead of a raw method exists, we capture exceptions on `tags` (to handle L5.1.28) and also examine the store if a `getStore` method is available (for older releases).