(#1622) Properly documented the public JWTAuth\Manager methods JWTAuth exposes.
When running phpstan/phpstan on a project which uses JWTAuth, errors are dutifully reported because JWTAuth also exposes all of the public methods of `JWTAuth\Manager`, without properly documenting in the class docblock that it does so.

```php
$token = JWTAuth::decode(JWTAuth::getToken())->toArray();
```
**PHP Stan reports it as an error:**
```
 ------ -----------------------------------------------------------------------------
  Line   app/Http/Controllers/AuthenticateController.php
 ------ -----------------------------------------------------------------------------
  165    Call to an undefined static method Tymon\JWTAuth\Facades\JWTAuth::decode().
 ------ -----------------------------------------------------------------------------
```
![image](https://user-images.githubusercontent.com/1125541/42651531-43f6450e-85d5-11e8-989a-1f8cdc454ee4.png)

Properly document that JWTAuth\JWTAuth exposes every public function of JWTAuth\Manager and the two traits it uses.