Without realpath this will work
Currently working on a project (with Laravel) that requires to be bundled in a Phar. The realpath is [not working within the Phar setup](https://www.php.net/manual/en/function.realpath.php#refsect1-function.realpath-returnvalues). 
This change will allow to work properly even with using a the relative path.