refactor!: eliminate usages of deprecated lcobucci/jwt methods
BREAKING CHANGE: The signature of `Tymon\JWTAuth\Providers\JWT\Lcobucci::__construct()` has been changed, because the previously injected dependencies are deprecated now. But it looks like it has only really been set up the way it was to facilitate testing, and the change should not affect users using it through the ServiceProvider.